Work: Finish implementation with instructions
Instead of generating code to do it, we generate an array of
instructions which in theory should be evaluated more efficiently.

However, the big problem is, that I can't convert a struct into a tuple.
Currently, I am trying to take the address of an object at comptime
which only exists at runtime.  That can not work.

However, the compiler does not seem to complain.  It produces the
incorrect result though.  I suspect that the parsing isn't working
correctly and thus we never trigger the error case.